pyproject.toml
[tool.poetry]
name = "rss-app-server"
version = "0.1.0"
description = ""
authors = ["jiazifa"]
readme = "README.md"
packages = [{include = "rss_app_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
pydantic = {extras = ["email"], version = "^2.0.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.0.2"
sqlalchemy = "^2.0.19"
orjson = "^3.9.2"
alembic = "^1.11.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"
yapf = "^0.40.1"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
line_length = 88

[tool.yapf]
based_on_style = "google"
column_limit = 88
blank_line_before_nested_class_or_def = true
dedent_closing_brackets = true
coalesce_brackets = false
spaces_before_comment = 4
split_before_logical_operator = true
allow_multiline_dictionary_keys = false
allow_split_before_dict_value = false

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::ResourceWarning",
]

[tool.pyright]

reportPrivateUsage = "none"
reportUnusedClass = "none"
reportUnusedFunction = "none"
reportTypedDictNotRequiredAccess = "warning"

# 全局配置
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    '^.venv\.py$',
]
plugins = ["sqlalchemy.ext.declarative.mypy", "pydantic.mypy"]

[[tool.mypy.overrides]]
module = "app"
warn_return_any = false

[[tool.mypy.overrides]]
module = [
    "sqlalchemy",
    "sqlalchemy.ext.declarative",
    "sqlalchemy.orm.decl_api"
]
ignore_missing_imports = true
warn_unused_ignores = false

[tool.ruff]
line-length = 88

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[per-file-ignores]
"__init__.py" = ["E402", "F401"]